Model & endpoint updates
