+++ src/js/helpers/stateHelper.js
@@ -10,9 +10,12 @@ export const fetchStateOverview = (id, year) => {
     const source = CancelToken.source();
     return {
         promise: Axios.request({
-            url: `v2/recipient/state/${id}?year=${year}`,
+            url: `v2/recipient/state/${id}/`,
             baseURL: kGlobalConstants.API,
             method: 'get',
+            params: {
+                year
+            },
             cancelToken: source.token
         }),
         cancel() {
