+++ src/js/models/v2/state/BaseStateProfile.js
@@ -3,64 +3,52 @@
  * Created by Lizzie Salita 5/1/18
  */
 
-import { formatNumberWithPrecision, formatMoney } from 'helpers/moneyFormatter';
-import CoreRecipient from './CoreRecipient';
+import { formatNumberWithPrecision, formatMoney, calculateUnitForSingleValue, formatMoneyWithPrecision } from 'helpers/moneyFormatter';
 
-const BaseStateProfile = Object.create(CoreRecipient);
-
-BaseStateProfile.populate = function populate(data) {
-    // reformat fields required by CoreRecipient
-    const coreData = {
-        name: data.name,
-        id: data.fips,
-        totalAmount: data.total_prime_amount,
-        totalAwards: data.total_prime_awards
-    };
-    this.populateCore(coreData);
-
-    this.type = data.type || '';
-    this.populationSourceYear = (data.pop_year && `${data.pop_year}`) || '';
-    this.incomeSourceYear = (data.mhi_year && `${data.mhi_year}`) || '';
-    this._population = data.population || 0;
-    this._awardAmountPerCapita = data.award_amount_per_capita || 0;
-    this._medianHouseholdIncome = data.median_household_income || 0;
-};
-
-// getter functions
-Object.defineProperty(BaseStateProfile, 'population', {
-    get() {
+const BaseStateProfile = {
+    populate(data) {
+        this.id = data.fips || null;
+        this.name = data.name || '';
+        this._totalAmount = parseFloat(data.total_prime_amount) || 0;
+        this._totalAwards = parseFloat(data.total_prime_awards) || 0;
+        this.type = data.type || '';
+        this.populationSourceYear = (data.pop_year && `${data.pop_year}`) || '';
+        this.incomeSourceYear = (data.mhi_year && `${data.mhi_year}`) || '';
+        this._population = data.population || 0;
+        this._awardAmountPerCapita = data.award_amount_per_capita || 0;
+        this._medianHouseholdIncome = data.median_household_income || 0;
+    },
+    get totalAmount() {
+        const units = calculateUnitForSingleValue(this._totalAmount);
+        return `${formatMoneyWithPrecision(this._totalAmount / units.unit, 1)} ${units.longLabel}`;
+    },
+    get totalAwards() {
+        return formatNumberWithPrecision(this._totalAwards, 0);
+    },
+    get population() {
         if (this._population === 0) {
             return '--';
         }
         return formatNumberWithPrecision(this._population, 0);
-    }
-});
-Object.defineProperty(BaseStateProfile, 'awardAmountPerCapita', {
-    get() {
+    },
+    get awardAmountPerCapita() {
         if (this._awardAmountPerCapita === 0) {
             return '--';
         }
         return formatMoney(this._awardAmountPerCapita);
-    }
-});
-Object.defineProperty(BaseStateProfile, 'medianHouseholdIncome', {
-    get() {
+    },
+    get medianHouseholdIncome() {
         if (this._medianHouseholdIncome === 0) {
             return '--';
         }
         return formatMoney(this._medianHouseholdIncome);
-    }
-});
-Object.defineProperty(BaseStateProfile, 'flag', {
-    get() {
+    },
+    get flag() {
         if (this.id) {
-            if (this._id < 10) {
-                return `https://s3-us-gov-west-1.amazonaws.com/da-public-files/usaspending_state_flags/0${this.id}.png`;
-            }
             return `https://s3-us-gov-west-1.amazonaws.com/da-public-files/usaspending_state_flags/${this.id}.png`;
         }
         return '';
     }
-});
+};
 
 export default BaseStateProfile;
